Give the flexbox a hint
Drew reveals the secret: if you hint to flexbox at the flex-item level
that this one might scroll inside, then it lets the flexbox fit in
the box around it.

This makes a difference when there's errors in an analysis, now it doesn't
grow to the right and off the page.